kontrola czy user usuwa swoja rezerwacje

+++ app/Http/Controllers/UserUpcomingReservationsController.php
@@ -19,7 +19,6 @@ class UserUpcomingReservationsController extends Controller
         $date->toDateString();
         //pobranie id obecnie zalogowanego użytkownika
         $logged_user_id = Auth::id();
-        echo($logged_user_id);
         $reservations = DB::table('reservations')
                             ->join('users', 'reservations.user_id', '=', 'users.id')
                             ->where('from', '>=', $date)
@@ -32,16 +31,23 @@ class UserUpcomingReservationsController extends Controller
         $date = Carbon::now();
         $date->toDateTimeString();
         $date->toDateString();
+        $logged_user_id = Auth::id();
         //pluck zeby wydobyc w arrayu bez {from: "xyz"}
         $reservation_start = DB::table('reservations')
                                 ->where('reservation_id', '=', $id)
                                 ->pluck('from');
         //obliczenie roznicy dni pomiedzy obecna data a poczatkiem rezerwacji
         $day_difference = $date->diffInDays($reservation_start[0]);
-        echo($day_difference);
+        //sprawdzenie czy rezerwacja nalezy do uzytkownika
+        $owner = DB::table('reservations')
+                    ->where('reservation_id', '=', $id)
+                    ->pluck('user_id');
+        if($owner[0] != $logged_user_id){
+            return redirect()->back() ->withInput()->withErrors(['not_authorized' => 'Nie masz do tego uprawnień!']);
+        }
         //sprawdzenie czy od obecnego momentu a poczatku rezerwacji zostalo mniej niz 3 dni
         if($day_difference <= 3){
-            return redirect()->back() ->withInput()->withErrors(['late' => 'Rezerwację można anulować najpóźniej 3 dni przed jej datą startu !']);
+            return redirect()->back() ->withInput()->withErrors(['late' => 'Rezerwację można anulować najpóźniej 3 dni przed jej datą startu!']);
         }
         DB::delete('delete from reservations where reservation_id = ?',[$id]);
         return redirect()->back()->with('message', 'Pomyślnie anulowano rezerwację !');
